src/mcp_servers/filesystem_mcp/requirements.txt
fastmcp
uv
mcpo
PyPDF2